initial setup,osm & frag comm

+++ app/src/main/java/com/example/weatherapp/model/pojo/WeatherResponse.kt
@@ -23,7 +23,7 @@ data class WeatherList(
     val clouds: Clouds,
     val wind: Wind,
     val visibility: Long,
-    val pop: Long,
+    val pop: Double,
     val sys: Sys,
     @SerializedName("dt_txt")
     val dtTxt: String,
